+++ crates/database/src/lib.rs
@@ -7,6 +7,8 @@ extern crate alloc as std;
 #[cfg(feature = "alloydb")]
 mod alloydb;
 
+pub use database_interface::*;
+
 pub mod in_memory_db;
 pub mod states;
