+++ crates/handler/src/lib.rs
@@ -7,6 +7,7 @@ extern crate alloc as std;
 
 // Mainnet related handlers.
 
+pub mod evm;
 pub mod execution;
 mod frame;
 mod frame_data;
@@ -20,6 +21,7 @@ mod precompile_provider;
 pub mod validation;
 
 // Public exports
+pub use evm::{ExecuteCommitEvm, ExecuteEvm};
 pub use frame::{return_create, return_eofcreate, ContextTrDbError, EthFrame, Frame};
 pub use frame_data::{FrameData, FrameResult};
 pub use handler::{EvmTr, EvmTrError, Handler};
