+++ crates/inspector/src/gas.rs
@@ -1,5 +1,5 @@
 //! GasIspector. Helper Inspector to calculate gas for others.
-use revm::interpreter::{CallOutcome, CreateOutcome, Gas};
+use interpreter::{CallOutcome, CreateOutcome, Gas};
 
 /// Helper that keeps track of gas.
 #[allow(dead_code)]
